API: Process images and keep the source file

+++ api/controllers/documentController.php
@@ -8,7 +8,7 @@ defined('EXEC') or die('Config not loaded');
  *
  * @author Niels Witte
  * @version 0.2
- * @date April 1st, 2014
+ * @date April 2nd, 2014
  * @since March 10th, 2014
  */
 class DocumentController {
@@ -17,7 +17,7 @@ class DocumentController {
     /**
      * Constructs a new controller with the given presentation
      *
-     * @param \Models\Presentation $presentation
+     * @param \Models\Document $document
      */
     public function __construct(\Models\Document $document = NULL) {
         $this->document = $document;
@@ -31,13 +31,33 @@ class DocumentController {
      */
     public function removeDocument() {
         $db = \Helper::getDB();
-        // First slides before parent document!
-        $db->where('documentId', $db->escape($this->document->getId()));
-        $db->delete('document_slides');
+        // Type is not available?
+        if($this->document->getType() == '') {
+            $this->document->getInfoFromDatabase();
+        }
 
+        // Remove comments for this document
+        $commentCtrl = new \Controllers\CommentController();
+        $commentCtrl->removeCommentsByItem($this->document->getType(), $this->document->getId());
+
+        // If it's a presentation, remove all comments for slides as well
+        if($this->document->getType() == 'presentation') {
+            $presentation = new \Models\Presentation($this->document->getId());
+
+            foreach($presentation->getSlides() as $slide) {
+                $commentCtrl->removeCommentsByItem('slide', $slide->getId());
+            }
+
+            // First slides before parent document!
+            $db->where('documentId', $db->escape($this->document->getId()));
+            $db->delete('document_slides');
+        }
+
+        // Remove the document itself
         $db->where('id', $db->escape($this->document->getId()));
         $result = $db->delete('documents');
 
+        // Something went wrong?
         if($result === FALSE) {
             throw new \Exception('Given Document ('. $this->document->getId() .') could not be removed from the CMS. The document is probably being used in a meeting.', 1);
         } else {
@@ -85,21 +105,28 @@ class DocumentController {
             'title'         => $db->escape($parameters['title']),
             'type'          => $db->escape($parameters['type']),
             'ownerId'       => $db->escape(\Auth::getUser()->getId()),
-            'creationDate'  => $db->escape(date('Y-m-d H:m:s'))
+            'creationDate'  => $db->escape(date('Y-m-d H:m:s')),
+            'file'          => $db->escape('source.'. $extension)
         );
         $fileId = $db->insert('documents', $data);
         // Insert successful?
         if($fileId !== FALSE) {
-            $filename = \Helper::saveFile($fileId .'.'. $extension, TEMP_LOCATION, $file);
-
+            $filename   = \Helper::saveFile($fileId .'.'. $extension, TEMP_LOCATION, $file);
             // File saved successful to temp?
             if($filename !== FALSE && file_exists($filename)) {
+                $targetDir  = FILES_LOCATION . DS . $parameters['type'] . DS . $fileId;
+
                 // Finally process images
-                if($data['type'] == 'image') {
-                    $destination = FILES_LOCATION . DS .'images'. DS . $fileId . DS . $fileId .'.'. $extension;
+                if($data['type'] == 'image' && in_array($extension, array('png', 'jpg', 'jpeg', 'gif'))) {
+                    $destination = $targetDir . DS . $fileId .'.'. $extension;
                     $result = \Helper::imageResize($filename, $destination);
                 } else {
+                    $result = TRUE;
+                }
 
+                // Move the file to target directory
+                if($result) {
+                    \Helper::moveFile($filename, $targetDir . DS .'source.'. $extension);
                 }
             }
         }
@@ -123,7 +150,7 @@ class DocumentController {
             }
         }
 
-        return $result;
+        return $result !== FALSE ? $fileId : $result;
     }
 
     /**
